# backend/docs/swagger.yaml
definitions:
  github_com_lshc_contract-hub_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_handler.SaveVersionRequest:
    properties:
      change_summary:
        type: string
      document_content:
        additionalProperties: true
        type: object
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 心智协同合同协作系统 V1 后端接口文档
  title: 心智协同合同协作系统 API
  version: "1.0"
paths:
...
  /api/contracts:
    get:
      consumes:
      - application/json
      description: 分页查询当前登录用户的合同列表，支持按合同名称/编号、状态、客户名称筛选
      parameters:
      - default: 1
        description: 页码，默认 1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认 10，最大 100
        in: query
        name: page_size
        type: integer
      - description: 合同名称或合同编号关键字
        in: query
        name: keyword
        type: string
      - description: 合同状态：0已导入 1已分享 2协作中 3已确认 4已完成 5已取消
        in: query
        name: status
        type: integer
      - description: 客户名称
        in: query
        name: customer_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页合同列表
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 合同列表
      tags:
      - 合同
    post:
      consumes:
      - multipart/form-data
      description: 上传 DOCX 文件创建合同，文件保存到阿里云 OSS，并自动生成 V1 合同版本
      parameters:
      - description: 合同名称
        in: formData
        name: contract_name
        required: true
        type: string
      - description: DOCX 文件，仅支持 .docx
        in: formData
        name: file
        required: true
        type: file
      - description: 合同备注
        in: formData
        name: description
        type: string
      - description: 客户名称/公司名称
        in: formData
        name: customer_name
        type: string
      - description: 客户联系人
        in: formData
        name: customer_contact
        type: string
      - description: 客户联系电话
        in: formData
        name: customer_phone
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 参数错误或文件不合法
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 创建合同
      tags:
      - 合同
  /api/contracts/{id}:
    get:
      consumes:
      - application/json
      description: 获取当前用户创建的合同详情，包含当前版本信息
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 合同详情
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在或无权访问
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 合同详情
      tags:
      - 合同
  /api/contracts/{id}/changes:
    get:
      consumes:
      - application/json
      description: 分页查询合同版本之间的变更记录
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码，默认 1
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量，默认 20，最大 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页变更记录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在或无权访问
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 变更记录列表
      tags:
      - 合同版本
...
  /api/contracts/{id}/versions:
    get:
      consumes:
      - application/json
      description: 分页查询合同版本列表
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码，默认 1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认 10，最大 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页版本列表
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在或无权访问
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 版本列表
      tags:
      - 合同版本
    post:
      consumes:
      - application/json
      description: 提交编辑后的结构化文档，自动生成新版本并记录变更
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - description: 新版本内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SaveVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在或无权访问
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "422":
          description: 文档对比失败
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 保存新版本
      tags:
      - 合同版本
  /api/contracts/{id}/versions/{versionId}:
    get:
      consumes:
      - application/json
      description: 获取指定历史版本的完整结构化内容
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - description: 版本ID
        in: path
        name: versionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 版本详情
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在、无权访问或版本不存在
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 版本详情
      tags:
      - 合同版本
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"